Split up the ItemRegistry for readability reasons.
